Move the client.ts code into a module structure
