+++ src/client/client-config.ts
@@ -1,0 +1,5 @@
+export type ClientConfig = {
+  deviceAddress: string;
+  bridgeAddress: string;
+  claimCode: string;
+};
